+++ src/adapter.ts
@@ -1,30 +1,22 @@
 //-----------------------------------------------------------------------------
-// vscode-catch2-test-adapter was written by Mate Pek, and is placed in the public
-// domain. The author hereby disclaims copyright to this source code.
+// vscode-catch2-test-adapter was written by Mate Pek, and is placed in the
+// public domain. The author hereby disclaims copyright to this source code.
 
-import { execFile } from "child_process";
-import * as path from "path";
-import * as vscode from "vscode";
-import * as fs from "fs";
-import {
-  TestAdapter,
-  TestEvent,
-  TestSuiteEvent,
-  TestLoadStartedEvent,
-  TestLoadFinishedEvent,
-  TestRunStartedEvent,
-  TestRunFinishedEvent
-} from "vscode-test-adapter-api";
-import * as util from "vscode-test-adapter-util";
-import { C2AllTestSuiteInfo, C2TestInfo, C2TestSuiteInfo, C2TestSuiteInfoBase } from "./Catch2";
+import {execFile} from 'child_process';
+import * as fs from 'fs';
+import * as path from 'path';
+import * as vscode from 'vscode';
+import {TestAdapter, TestEvent, TestLoadFinishedEvent, TestLoadStartedEvent, TestRunFinishedEvent, TestRunStartedEvent, TestSuiteEvent} from 'vscode-test-adapter-api';
+import * as util from 'vscode-test-adapter-util';
+
+import {C2AllTestSuiteInfo, C2TestInfo} from './catch2';
 
 export class Catch2TestAdapter implements TestAdapter, vscode.Disposable {
-  private readonly testsEmitter = new vscode.EventEmitter<
-    TestLoadStartedEvent | TestLoadFinishedEvent
-  >();
-  readonly testStatesEmitter = new vscode.EventEmitter<
-    TestRunStartedEvent | TestRunFinishedEvent | TestSuiteEvent | TestEvent
-  >();
+  private readonly testsEmitter =
+      new vscode.EventEmitter<TestLoadStartedEvent|TestLoadFinishedEvent>();
+  readonly testStatesEmitter =
+      new vscode.EventEmitter<TestRunStartedEvent|TestRunFinishedEvent|
+                              TestSuiteEvent|TestEvent>();
   private readonly autorunEmitter = new vscode.EventEmitter<void>();
 
   private readonly watchers: Map<string, fs.FSWatcher> = new Map();
@@ -35,63 +27,40 @@ export class Catch2TestAdapter implements TestAdapter, vscode.Disposable {
   private readonly disposables: Array<vscode.Disposable> = new Array();
 
   private isEnabledSourceDecoration = true;
-  private readonly variableResolvedPair: [string, string][] = [
-    ["${workspaceFolder}", this.workspaceFolder.uri.fsPath]
-  ];
+  private readonly variableResolvedPair: [string, string][] =
+      [['${workspaceFolder}', this.workspaceFolder.uri.fsPath]];
 
   getIsEnabledSourceDecoration(): boolean {
     return this.isEnabledSourceDecoration;
   }
 
   constructor(
-    public readonly workspaceFolder: vscode.WorkspaceFolder,
-    public readonly log: util.Log
-  ) {
+      public readonly workspaceFolder: vscode.WorkspaceFolder,
+      public readonly log: util.Log) {
     this.disposables.push(
-      vscode.workspace.onDidChangeConfiguration(configChange => {
-        if (
-          configChange.affectsConfiguration(
-            "catch2TestExplorer.defaultGroupFileLevelRun",
-            this.workspaceFolder.uri
-          ) ||
-          configChange.affectsConfiguration(
-            "catch2TestExplorer.defaultEnv",
-            this.workspaceFolder.uri
-          ) ||
-          configChange.affectsConfiguration(
-            "catch2TestExplorer.defaultCwd",
-            this.workspaceFolder.uri
-          ) ||
-          configChange.affectsConfiguration(
-            "catch2TestExplorer.defaultWorkerMaxNumberPerFile",
-            this.workspaceFolder.uri
-          ) ||
-          configChange.affectsConfiguration(
-            "catch2TestExplorer.globalWorkerMaxNumber",
-            this.workspaceFolder.uri
-          ) ||
-          configChange.affectsConfiguration(
-            "catch2TestExplorer.executables",
-            this.workspaceFolder.uri
-          )
-        ) {
-          this.load();
-        }
-      })
-    );
+        vscode.workspace.onDidChangeConfiguration(configChange => {
+          if (configChange.affectsConfiguration(
+                  'catch2TestExplorer.defaultEnv', this.workspaceFolder.uri) ||
+              configChange.affectsConfiguration(
+                  'catch2TestExplorer.defaultCwd', this.workspaceFolder.uri) ||
+              configChange.affectsConfiguration(
+                  'catch2TestExplorer.workerMaxNumber',
+                  this.workspaceFolder.uri) ||
+              configChange.affectsConfiguration(
+                  'catch2TestExplorer.executables', this.workspaceFolder.uri)) {
+            this.load();
+          }
+        }));
 
     this.disposables.push(
-      vscode.workspace.onDidChangeConfiguration(configChange => {
-        if (
-          configChange.affectsConfiguration(
-            "catch2TestExplorer.enableSourceDecoration",
-            this.workspaceFolder.uri
-          )
-        ) {
-          this.isEnabledSourceDecoration = this.getEnableSourceDecoration(this.getConfiguration());
-        }
-      })
-    );
+        vscode.workspace.onDidChangeConfiguration(configChange => {
+          if (configChange.affectsConfiguration(
+                  'catch2TestExplorer.enableSourceDecoration',
+                  this.workspaceFolder.uri)) {
+            this.isEnabledSourceDecoration =
+                this.getEnableSourceDecoration(this.getConfiguration());
+          }
+        }));
 
     this.allTests = new C2AllTestSuiteInfo(this, 1);
   }
@@ -100,16 +69,16 @@ export class Catch2TestAdapter implements TestAdapter, vscode.Disposable {
     this.disposables.forEach(d => {
       d.dispose();
     });
-    while (this.disposables.shift() !== undefined);
+    while (this.disposables.shift() !== undefined)
+      ;
   }
 
-  get testStates(): vscode.Event<
-    TestRunStartedEvent | TestRunFinishedEvent | TestSuiteEvent | TestEvent
-  > {
+  get testStates(): vscode.Event<TestRunStartedEvent|TestRunFinishedEvent|
+                                 TestSuiteEvent|TestEvent> {
     return this.testStatesEmitter.event;
   }
 
-  get tests(): vscode.Event<TestLoadStartedEvent | TestLoadFinishedEvent> {
+  get tests(): vscode.Event<TestLoadStartedEvent|TestLoadFinishedEvent> {
     return this.testsEmitter.event;
   }
 
@@ -117,72 +86,103 @@ export class Catch2TestAdapter implements TestAdapter, vscode.Disposable {
     return this.autorunEmitter.event;
   }
 
-  loadSuite(
-    exe: ExecutableConfig,
-    parentSuite: C2TestSuiteInfoBase,
-    oldSuite: C2TestSuiteInfo | undefined
-  ): Promise<void> {
+  loadSuite(exe: ExecutableConfig): Promise<void> {
     return new Promise<void>((resolve, reject) => {
       try {
         execFile(
-          exe.path,
-          ["--verbosity", "high", "--list-test-names-only"],
-          (error: Error | null, stdout: string, stderr: string) => {
-            const suite = parentSuite.createChildSuite(
-              exe.name,
-              exe.workerMaxNumber,
-              this.getDefaultGroupFileLevelRun(this.getConfiguration()),
-              exe.path,
-              { cwd: exe.cwd, env: exe.env },
-              oldSuite
-            );
+            exe.path,
+            [
+              '[.],*', '--verbosity', 'high', '--list-tests', '--use-colour',
+              'no'
+            ],
+            (error: Error|null, stdout: string, stderr: string) => {
+              const suite = this.allTests.createChildSuite(
+                  exe.name, exe.path, {cwd: exe.cwd, env: exe.env});
 
-            let lines = stdout.split(/[\n\r]+/);
-            for (var line of lines) {
-              const match = line.match(/^([^@]+)@([^\(]+)\(([0-9]+)\)$/);
-              if (match && match.length > 0) {
-                const filePath = exe.cwd ? path.resolve(exe.cwd, match[2]) : match[2];
-                suite.createChildTest(match[1].trim(), filePath, Number(match[3]));
+              let lines = stdout.split(/\n\r?/);
+
+              if (lines.length == 0) this.log.error('Empty test list.');
+
+              while (lines[lines.length - 1].length == 0) lines.pop();
+
+              let i = 1;
+              while (i < lines.length - 1) {
+                if (lines[i][0] != ' ')
+                  this.log.error(
+                      'Wrong test list output format: ' + lines.toString());
+
+                const testName = lines[i++].substr(2);
+
+                let filePath = '';
+                let line = 0;
+                {
+                  const fileLine = lines[i++].substr(4);
+                  const colon = fileLine.indexOf(':');
+                  filePath = fileLine.substring(0, colon);
+                  if (exe.cwd) filePath = path.resolve(exe.cwd, filePath);
+                  line = Number(fileLine.substring(colon + 1));
+                }
+
+                let description = lines[i++].substr(4);
+                if (description.startsWith('(NO DESCRIPTION)'))
+                  description = '';
+
+                let tags: string[] = [];
+                if (lines[i].length > 6 && lines[i][6] === '[') {
+                  tags = lines[i].trim().split(']');
+                  tags.pop();
+                  for (let j = 0; j < tags.length; ++j) tags[j] += ']';
+                  ++i;
+                }
+
+                suite.createChildTest(
+                    testName, description, tags, filePath, line);
               }
-            }
 
-            let watcher = this.watchers.get(exe.path);
+              let watcher = this.watchers.get(exe.path);
 
-            if (watcher != undefined) {
-              watcher.close();
-            }
+              if (watcher != undefined) {
+                watcher.close();
+              }
 
-            watcher = fs.watch(exe.path);
-            this.watchers.set(exe.path, watcher);
+              watcher = fs.watch(exe.path);
+              this.watchers.set(exe.path, watcher);
 
-            watcher.on("change", (eventType: string, filename: string) => {
-              if (!fs.existsSync(exe.path)) {
-                watcher!.close();
-                this.watchers.delete(exe.path);
-                this.testsEmitter.fire({ type: "started" });
-                parentSuite.removeChild(suite);
-                this.testsEmitter.fire({
-                  type: "finished",
-                  suite: this.allTests
-                });
-              } else if (this.isDebugging) {
+              watcher.on('change', (eventType: string, filename: string) => {
+                // need some time here:
+                const waitAndThenTry =
+                    (remainingIteration: number, delay: number) => {
+                      if (remainingIteration == 0) return;
+                      if (!fs.existsSync(exe.path)) {
+                        setTimeout(waitAndThenTry, delay, [
+                          remainingIteration - 1, Math.max(delay * 2, 2000)
+                        ]);
+                      } else {
+                        this.testsEmitter.fire({type: 'started'});
+                        this.loadSuite(exe).then(() => {
+                          this.testsEmitter.fire(
+                              {type: 'finished', suite: this.allTests});
+                        });
+                      }
+                    };
+
                 // change event can arrive during debug session on osx (why?)
-              } else {
-                this.testsEmitter.fire({ type: "started" });
-                this.loadSuite(exe, parentSuite, suite).then(() => {
-                  this.testsEmitter.fire({
-                    type: "finished",
-                    suite: this.allTests
-                  });
-                });
-              }
-            });
+                if (!this.isDebugging) {
+                  watcher!.close();
+                  this.watchers.delete(exe.path);
+                  this.testsEmitter.fire({type: 'started'});
+                  this.allTests.removeChild(suite);
+                  this.testsEmitter.fire(
+                      {type: 'finished', suite: this.allTests});
 
-            resolve();
-          }
-        );
+                  waitAndThenTry(10, 128);
+                }
+              });
+
+              resolve();
+            });
       } catch (e) {
-        console.error("Something is wrong.", e);
+        console.error('Something is wrong.', e);
         resolve();
       }
     });
@@ -191,7 +191,7 @@ export class Catch2TestAdapter implements TestAdapter, vscode.Disposable {
   load(): Promise<void> {
     this.cancel();
 
-    this.testsEmitter.fire({ type: "started" });
+    this.testsEmitter.fire({type: 'started'});
 
     this.watchers.forEach((value, key) => {
       value.close();
@@ -200,26 +200,26 @@ export class Catch2TestAdapter implements TestAdapter, vscode.Disposable {
 
     const config = this.getConfiguration();
 
-    const allTests = new C2AllTestSuiteInfo(this, this.getGlobalWorkerMaxNumber(config));
+    this.allTests =
+        new C2AllTestSuiteInfo(this, this.getWorkerMaxNumber(config));
 
     const executables = this.getExecutables(config);
 
     return executables
-      .then((execs: ExecutableConfig[]) => {
-        let testListReaders = Promise.resolve();
+        .then((execs: ExecutableConfig[]) => {
+          let testListReaders = Promise.resolve();
 
-        execs.forEach(exe => {
-          testListReaders = testListReaders.then(() => {
-            return this.loadSuite(exe, allTests, undefined);
+          execs.forEach(exe => {
+            testListReaders = testListReaders.then(() => {
+              return this.loadSuite(exe);
+            });
           });
-        });
 
-        return testListReaders;
-      })
-      .then(() => {
-        this.allTests = allTests;
-        this.testsEmitter.fire({ type: "finished", suite: allTests });
-      });
+          return testListReaders;
+        })
+        .then(() => {
+          this.testsEmitter.fire({type: 'finished', suite: this.allTests});
+        });
   }
 
   cancel(): void {
@@ -228,7 +228,7 @@ export class Catch2TestAdapter implements TestAdapter, vscode.Disposable {
 
   run(tests: string[]): Promise<void> {
     if (this.isDebugging) {
-      throw "Catch2: Test is currently being debugged.";
+      throw 'Catch2: Test is currently being debugged.';
     }
 
     if (this.isRunning == 0) {
@@ -236,19 +236,19 @@ export class Catch2TestAdapter implements TestAdapter, vscode.Disposable {
       const always = () => {
         this.isRunning -= 1;
       };
-      return this.allTests.test(tests).then(always, always);
+      return this.allTests.run(tests).then(always, always);
     }
 
-    throw Error("Catch2 Test Adapter: Test(s) are currently being run.");
+    throw Error('Catch2 Test Adapter: Test(s) are currently being run.');
   }
 
   async debug(tests: string[]): Promise<void> {
     if (this.isDebugging) {
-      throw "Catch2: Test is currently being debugged.";
+      throw 'Catch2: Test is currently being debugged.';
     }
 
     if (this.isRunning > 0) {
-      throw "Catch2: Test(s) are currently being run.";
+      throw 'Catch2: Test(s) are currently being run.';
     }
 
     console.assert(tests.length === 1);
@@ -256,26 +256,26 @@ export class Catch2TestAdapter implements TestAdapter, vscode.Disposable {
     console.assert(info !== undefined);
 
     if (!(info instanceof C2TestInfo)) {
-      throw "Can't choose a group, only a single test.";
+      throw 'Can\'t choose a group, only a single test.';
     }
 
     const testInfo = <C2TestInfo>info;
 
     const getDebugConfiguration = (): vscode.DebugConfiguration => {
       const debug: vscode.DebugConfiguration = {
-        name: "Catch2: " + testInfo.label,
-        request: "launch",
-        type: ""
+        name: 'Catch2: ' + testInfo.label,
+        request: 'launch',
+        type: ''
       };
 
-      const template = this.getDebugConfigurationTemplate(this.getConfiguration());
+      const template =
+          this.getDebugConfigurationTemplate(this.getConfiguration());
       let resolveDebugVariables: [string, any][] = this.variableResolvedPair;
       resolveDebugVariables = resolveDebugVariables.concat([
-        ["${label}", testInfo.label],
-        ["${exec}", testInfo.execPath],
-        ["${args}", testInfo.execParams.slice(0, 1).concat("--reporter", "console")],
-        ["${cwd}", testInfo.execOptions.cwd!],
-        ["${envObj}", testInfo.execOptions.env!]
+        ['${label}', testInfo.label], ['${exec}', testInfo.parent.execPath],
+        ['${args}', [testInfo.getEscapedTestName(), '--reporter', 'console']],
+        ['${cwd}', testInfo.parent.execOptions.cwd!],
+        ['${envObj}', testInfo.parent.execOptions.env!]
       ]);
 
       if (template !== null) {
@@ -288,31 +288,29 @@ export class Catch2TestAdapter implements TestAdapter, vscode.Disposable {
 
         return debug;
       } else {
-        //lets try to recognise existing extensions
-        vscode.extensions.getExtension("");
+        // lets try to recognise existing extensions
+        vscode.extensions.getExtension('');
       }
 
-      throw "Catch2: For debug 'debugConfigurationTemplate' should be set.";
+      throw 'Catch2: For debug \'debugConfigurationTemplate\' should be set.';
     };
 
     const debugConfig = getDebugConfiguration();
 
     this.isDebugging = true;
 
-    const debugSessionStarted = await vscode.debug.startDebugging(
-      this.workspaceFolder,
-      debugConfig
-    );
+    const debugSessionStarted =
+        await vscode.debug.startDebugging(this.workspaceFolder, debugConfig);
 
     if (!debugSessionStarted) {
-      console.error("Failed starting the debug session - aborting");
+      console.error('Failed starting the debug session - aborting');
       this.isDebugging = false;
       return;
     }
 
     const currentSession = vscode.debug.activeDebugSession;
     if (!currentSession) {
-      console.error("No active debug session - aborting");
+      console.error('No active debug session - aborting');
       this.isDebugging = false;
       return;
     }
@@ -324,7 +322,7 @@ export class Catch2TestAdapter implements TestAdapter, vscode.Disposable {
     await new Promise<void>((resolve, reject) => {
       const subscription = vscode.debug.onDidTerminateDebugSession(session => {
         if (currentSession != session) return;
-        console.info("Debug session ended");
+        console.info('Debug session ended');
         resolve();
         subscription.dispose();
       });
@@ -332,47 +330,45 @@ export class Catch2TestAdapter implements TestAdapter, vscode.Disposable {
   }
 
   private getConfiguration(): vscode.WorkspaceConfiguration {
-    return vscode.workspace.getConfiguration("catch2TestExplorer", this.workspaceFolder.uri);
+    return vscode.workspace.getConfiguration(
+        'catch2TestExplorer', this.workspaceFolder.uri);
   }
 
-  private getDebugConfigurationTemplate(
-    config: vscode.WorkspaceConfiguration
-  ): { [prop: string]: string } | null {
-    const o = config.get<any>("debugConfigurationTemplate", null);
+  private getDebugConfigurationTemplate(config: vscode.WorkspaceConfiguration):
+      {[prop: string]: string}|null {
+    const o = config.get<any>('debugConfigurationTemplate', null);
 
     if (o === null) return null;
 
-    const result: { [prop: string]: string } = {};
+    const result: {[prop: string]: string} = {};
 
     for (const prop in o) {
       const val = o[prop];
       if (val === undefined || val === null) {
         delete result.prop;
       } else {
-        result[prop] = this.resolveVariables(String(val), this.variableResolvedPair);
+        result[prop] =
+            this.resolveVariables(String(val), this.variableResolvedPair);
       }
     }
     return result;
   }
 
-  private getDefaultGroupFileLevelRun(config: vscode.WorkspaceConfiguration): boolean {
-    return config.get<boolean>("defaultGroupFileLevelRun", false);
-  }
-
   private getGlobalAndDefaultEnvironmentVariables(
-    config: vscode.WorkspaceConfiguration
-  ): { [prop: string]: string | undefined } {
+      config: vscode.WorkspaceConfiguration):
+      {[prop: string]: string|undefined} {
     const processEnv = process.env;
-    const configEnv: { [prop: string]: any } = config.get("defaultEnv") || {};
+    const configEnv: {[prop: string]: any} = config.get('defaultEnv') || {};
 
-    const resultEnv = { ...processEnv };
+    const resultEnv = {...processEnv};
 
     for (const prop in configEnv) {
       const val = configEnv[prop];
       if (val === undefined || val === null) {
         delete resultEnv.prop;
       } else {
-        resultEnv[prop] = this.resolveVariables(String(val), this.variableResolvedPair);
+        resultEnv[prop] =
+            this.resolveVariables(String(val), this.variableResolvedPair);
       }
     }
 
@@ -382,9 +378,7 @@ export class Catch2TestAdapter implements TestAdapter, vscode.Disposable {
   private getDefaultCwd(config: vscode.WorkspaceConfiguration): string {
     const dirname = this.workspaceFolder.uri.fsPath;
     const cwd = this.resolveVariables(
-      config.get<string>("defaultCwd", dirname),
-      this.variableResolvedPair
-    );
+        config.get<string>('defaultCwd', dirname), this.variableResolvedPair);
     if (path.isAbsolute(cwd)) {
       return cwd;
     } else {
@@ -392,107 +386,97 @@ export class Catch2TestAdapter implements TestAdapter, vscode.Disposable {
     }
   }
 
-  private getDefaultWorkerMaxNumberPerFile(config: vscode.WorkspaceConfiguration): number {
-    return config.get<number>("defaultWorkerMaxNumberPerFile", 1);
-  }
-
-  private getGlobalWorkerMaxNumber(config: vscode.WorkspaceConfiguration): number {
-    return config.get<number>("globalWorkerMaxNumber", 4);
+  private getWorkerMaxNumber(config: vscode.WorkspaceConfiguration): number {
+    return config.get<number>('workerMaxNumber', 4);
   }
 
   private getGlobalAndCurrentEnvironmentVariables(
-    config: vscode.WorkspaceConfiguration,
-    configEnv: { [prop: string]: any }
-  ): { [prop: string]: any } {
+      config: vscode.WorkspaceConfiguration,
+      configEnv: {[prop: string]: any}): {[prop: string]: any} {
     const processEnv = process.env;
-    const resultEnv = { ...processEnv };
+    const resultEnv = {...processEnv};
 
     for (const prop in configEnv) {
       const val = configEnv[prop];
       if (val === undefined || val === null) {
         delete resultEnv.prop;
       } else {
-        resultEnv[prop] = this.resolveVariables(String(val), this.variableResolvedPair);
+        resultEnv[prop] =
+            this.resolveVariables(String(val), this.variableResolvedPair);
       }
     }
 
     return resultEnv;
   }
 
-  private getEnableSourceDecoration(config: vscode.WorkspaceConfiguration): boolean {
-    return config.get<boolean>("enableSourceDecoration", true);
+  private getEnableSourceDecoration(config: vscode.WorkspaceConfiguration):
+      boolean {
+    return config.get<boolean>('enableSourceDecoration', true);
   }
 
-  private getExecutables(config: vscode.WorkspaceConfiguration): Promise<ExecutableConfig[]> {
+  private getExecutables(config: vscode.WorkspaceConfiguration):
+      Promise<ExecutableConfig[]> {
     const globalWorkingDirectory = this.getDefaultCwd(config);
 
     let executables: ExecutableConfig[] = [];
 
-    const configExecs:
-      | undefined
-      | string
-      | string[]
-      | { [prop: string]: any }
-      | { [prop: string]: any }[] = config.get("executables");
+    const configExecs:|undefined|string|string[]|{[prop: string]: any}|
+        {[prop: string]: any}[] = config.get('executables');
 
     const fullPath = (p: string): string => {
       return path.isAbsolute(p) ? p : this.resolveRelPath(p);
     };
 
     const addObject = (o: Object): void => {
-      const name: string = o.hasOwnProperty("name") ? (<any>o)["name"] : "${dirname} : ${name}";
-      if (!o.hasOwnProperty("path") || (<any>o)["path"] === null) {
-        console.warn(Error("'path' is a requireds property."));
+      const name: string =
+          o.hasOwnProperty('name') ? (<any>o)['name'] : '${dirname} : ${name}';
+      if (!o.hasOwnProperty('path') || (<any>o)['path'] === null) {
+        console.warn(Error('\'path\' is a requireds property.'));
         return;
       }
       const p: string = fullPath(
-        this.resolveVariables((<any>o)["path"], this.variableResolvedPair)
-      );
-      const regex: string = o.hasOwnProperty("regex") ? (<any>o)["regex"] : "";
-      const pool: number = o.hasOwnProperty("workerMaxNumber")
-        ? Number((<any>o)["workerMaxNumber"])
-        : this.getDefaultWorkerMaxNumberPerFile(config);
-      const cwd: string = o.hasOwnProperty("cwd") ? (<any>o)["cwd"] : globalWorkingDirectory;
-      const env: { [prop: string]: any } = o.hasOwnProperty("env")
-        ? this.getGlobalAndCurrentEnvironmentVariables(config, (<any>o)["env"])
-        : this.getGlobalAndDefaultEnvironmentVariables(config);
-      const regexRecursive: boolean = o.hasOwnProperty("recursiveRegex")
-        ? (<any>o)["recursiveRegex"]
-        : false;
+          this.resolveVariables((<any>o)['path'], this.variableResolvedPair));
+      const regex: string = o.hasOwnProperty('regex') ? (<any>o)['regex'] : '';
+      const cwd: string =
+          o.hasOwnProperty('cwd') ? (<any>o)['cwd'] : globalWorkingDirectory;
+      const env: {[prop: string]: any} = o.hasOwnProperty('env') ?
+          this.getGlobalAndCurrentEnvironmentVariables(
+              config, (<any>o)['env']) :
+          this.getGlobalAndDefaultEnvironmentVariables(config);
+      const regexRecursive: boolean = o.hasOwnProperty('recursiveRegex') ?
+          (<any>o)['recursiveRegex'] :
+          false;
 
       if (regex.length > 0) {
         const recursiveAdd = (directory: string): void => {
-          const children = fs.readdirSync(directory, "utf8");
+          const children = fs.readdirSync(directory, 'utf8');
           children.forEach(child => {
             const childPath = path.resolve(directory, child);
             const childStat = fs.statSync(childPath);
             if (childPath.match(regex) && childStat.isFile()) {
-              let resolvedName = name + " : " + child;
+              let resolvedName = name + ' : ' + child;
               let resolvedCwd = cwd;
               try {
                 resolvedName = this.resolveVariables(name, [
-                  ["${absDirname}", p],
-                  ["${dirname}", path.relative(this.workspaceFolder.uri.fsPath, p)],
-                  ["${name}", child]
+                  ['${absDirname}', p],
+                  [
+                    '${dirname}',
+                    path.relative(this.workspaceFolder.uri.fsPath, p)
+                  ],
+                  ['${name}', child]
                 ]);
                 resolvedCwd = this.resolveVariables(
-                  cwd,
-                  this.variableResolvedPair.concat([
-                    ["${absDirname}", p],
-                    ["${dirname}", path.relative(this.workspaceFolder.uri.fsPath, p)]
-                  ])
-                );
-              } catch (e) {}
-              executables.push(
-                new ExecutableConfig(
-                  resolvedName,
-                  childPath,
-                  regex,
-                  pool,
-                  fullPath(resolvedCwd),
-                  env
-                )
-              );
+                    cwd, this.variableResolvedPair.concat([
+                      ['${absDirname}', p],
+                      [
+                        '${dirname}',
+                        path.relative(this.workspaceFolder.uri.fsPath, p)
+                      ]
+                    ]));
+              } catch (e) {
+              }
+              executables.push(new ExecutableConfig(
+                  resolvedName, childPath, regex, fullPath(resolvedCwd), env));
             } else if (childStat.isDirectory() && regexRecursive) {
               recursiveAdd(childPath);
             }
@@ -503,7 +487,7 @@ export class Catch2TestAdapter implements TestAdapter, vscode.Disposable {
           if (stat.isDirectory()) {
             recursiveAdd(p);
           } else if (stat.isFile()) {
-            executables.push(new ExecutableConfig(name, p, regex, pool, cwd, env));
+            executables.push(new ExecutableConfig(name, p, regex, cwd, env));
           } else {
             // do nothing
           }
@@ -511,37 +495,28 @@ export class Catch2TestAdapter implements TestAdapter, vscode.Disposable {
           this.log.error(e.message);
         }
       } else {
-        executables.push(new ExecutableConfig(name, p, regex, pool, cwd, env));
+        executables.push(new ExecutableConfig(name, p, regex, cwd, env));
       }
     };
 
-    if (typeof configExecs === "string") {
+    if (typeof configExecs === 'string') {
       if (configExecs.length == 0) return Promise.resolve([]);
-      executables.push(
-        new ExecutableConfig(
+      executables.push(new ExecutableConfig(
           configExecs,
-          fullPath(this.resolveVariables(configExecs, this.variableResolvedPair)),
-          "",
-          1,
-          globalWorkingDirectory,
-          []
-        )
-      );
+          fullPath(
+              this.resolveVariables(configExecs, this.variableResolvedPair)),
+          '', globalWorkingDirectory, []));
     } else if (Array.isArray(configExecs)) {
       for (var i = 0; i < configExecs.length; ++i) {
-        if (typeof configExecs[i] == "string") {
+        if (typeof configExecs[i] == 'string') {
           const configExecsStr = String(configExecs[i]);
           if (configExecsStr.length > 0) {
-            executables.push(
-              new ExecutableConfig(
-                this.resolveVariables(configExecsStr, this.variableResolvedPair),
-                fullPath(this.resolveVariables(configExecsStr, this.variableResolvedPair)),
-                "",
-                1,
-                globalWorkingDirectory,
-                []
-              )
-            );
+            executables.push(new ExecutableConfig(
+                this.resolveVariables(
+                    configExecsStr, this.variableResolvedPair),
+                fullPath(this.resolveVariables(
+                    configExecsStr, this.variableResolvedPair)),
+                '', globalWorkingDirectory, []));
           }
         } else {
           addObject(configExecs[i]);
@@ -550,43 +525,40 @@ export class Catch2TestAdapter implements TestAdapter, vscode.Disposable {
     } else if (configExecs instanceof Object) {
       addObject(configExecs);
     } else {
-      throw "Catch2 config error: wrong type: executables";
+      throw 'Catch2 config error: wrong type: executables';
     }
 
-    executables.sort((a: ExecutableConfig, b: ExecutableConfig) => {
-      return a.name < b.name ? -1 : a.name > b.name ? 1 : 0;
-    });
     return this.filterVerifiedCatch2TestExecutables(executables);
   }
 
   private verifyIsCatch2TestExecutable(path: string): Promise<boolean> {
     return new Promise<boolean>((resolve, reject) => {
       try {
-        execFile(path, ["--help"], (error: Error | null, stdout: string, stderr: string) => {
-          if (stdout.indexOf("Catch v2.") != -1) {
-            resolve(true);
-          } else {
-            resolve(false);
-          }
-        });
+        execFile(
+            path, ['--help'],
+            (error: Error|null, stdout: string, stderr: string) => {
+              if (stdout.indexOf('Catch v2.') != -1) {
+                resolve(true);
+              } else {
+                resolve(false);
+              }
+            });
       } catch (e) {
         resolve(false);
       }
     });
   }
 
-  private filterVerifiedCatch2TestExecutables(
-    executables: ExecutableConfig[]
-  ): Promise<ExecutableConfig[]> {
+  private filterVerifiedCatch2TestExecutables(executables: ExecutableConfig[]):
+      Promise<ExecutableConfig[]> {
     const verified: ExecutableConfig[] = [];
     const promises: Promise<void>[] = [];
 
     executables.forEach(exec => {
-      promises.push(
-        this.verifyIsCatch2TestExecutable(exec.path).then((isCatch2: boolean) => {
-          if (isCatch2) verified.push(exec);
-        })
-      );
+      promises.push(this.verifyIsCatch2TestExecutable(exec.path).then(
+          (isCatch2: boolean) => {
+            if (isCatch2) verified.push(exec);
+          }));
     });
 
     return Promise.all(promises).then(() => {
@@ -595,9 +567,9 @@ export class Catch2TestAdapter implements TestAdapter, vscode.Disposable {
   }
 
   private resolveVariables(value: any, varValue: [string, any][]): any {
-    if (typeof value == "string") {
+    if (typeof value == 'string') {
       for (let i = 0; i < varValue.length; ++i) {
-        if (value === varValue[i][0] && typeof varValue[i][1] != "string") {
+        if (value === varValue[i][0] && typeof varValue[i][1] != 'string') {
           return varValue[i][1];
         }
         value = value.replace(varValue[i][0], varValue[i][1]);
@@ -605,7 +577,7 @@ export class Catch2TestAdapter implements TestAdapter, vscode.Disposable {
       return value;
     } else if (Array.isArray(value)) {
       return (<any[]>value).map((v: any) => this.resolveVariables(v, varValue));
-    } else if (typeof value == "object") {
+    } else if (typeof value == 'object') {
       const newValue: any = {};
       for (const prop in value) {
         newValue[prop] = this.resolveVariables(value[prop], varValue);
@@ -629,11 +601,7 @@ export class Catch2TestAdapter implements TestAdapter, vscode.Disposable {
 
 class ExecutableConfig {
   constructor(
-    public readonly name: string,
-    public readonly path: string,
-    public readonly regex: string,
-    public readonly workerMaxNumber: number,
-    public readonly cwd: string,
-    public readonly env: { [prop: string]: any }
-  ) {}
+      public readonly name: string, public readonly path: string,
+      public readonly regex: string, public readonly cwd: string,
+      public readonly env: {[prop: string]: any}) {}
 }